+++ controller/serviceService/deleteService.js
@@ -1,9 +1,13 @@
 const { Services } = require("../../schema");
 
 const deleteService = async (req, res) => {
-  const { id } = req.params;
-  const stat = await Services.findByIdAndDelete(id);
-  res.status(200).json(stat);
+  if (req.user) {
+    const { id } = req.params;
+    const stat = await Services.findByIdAndDelete(id);
+    res.status(200).json(stat);
+  } else {
+    res.status(401).json({ status: false, message: "unauthorized access" });
+  }
 };
 
 module.exports = deleteService;
